+++ inst/include/wk/rcpp-sexp-writer.hpp
@@ -92,7 +92,7 @@ protected:
   }
 
   void nextCoordinate(const WKGeometryMeta& meta, const WKCoord& coord, uint32_t coordId) {
-    for (int i = 0; i < coord.size(); i++) {
+    for (size_t i = 0; i < coord.size(); i++) {
       this->currentCoordinates(coordId, i) = coord[i];
     }
   }
